refactor: separate v1 and v2 biome config implementations
- Move filterFilesWithOrderedPatterns into biome2Config.ts (v2-only)
- Slim down biomeConfigUtils.ts to only truly shared functions:
  - loadRawBiomeConfigFile (JSON5 parsing)
  - getClosestBiomeConfigPath (find-up logic)
- Restore biomeConfig.ts to self-contained v1 implementation
- Fix microsyntax tests by renaming config files to standard names
  (biome.jsonc) and adding biome2 test data to ignore lists
- Add biome2 test data directory to biome ignore lists to prevent
  Biome 1.x from parsing Biome 2.x syntax
