+++ biome.jsonc
@@ -44,6 +44,8 @@
 
 			// Test json
 			"build-tools/packages/build-tools/src/test/data/biome/empty.jsonc",
+			// Biome 2.x test data - these configs use Biome 2.x syntax that Biome 1.x can't parse
+			"build-tools/packages/build-tools/src/test/data/biome2/**",
 			"experimental/dds/tree/src/test/documents/**",
 			"packages/dds/map/src/test/mocha/snapshots/**/*.json",
 			"packages/dds/matrix/src/test/results/**/*.json",
