+++ build-tools/packages/build-tools/src/common/biome2Config.ts
@@ -6,16 +6,13 @@
 import { strict as assert } from "node:assert/strict";
 import { stat } from "node:fs/promises";
 import path from "node:path";
+import ignore from "ignore";
+import multimatch from "multimatch";
 import { merge } from "ts-deepmerge";
 import type { Opaque } from "type-fest";
 
 import type { Configuration as Biome2ConfigRaw } from "./biome2ConfigTypes";
-import {
-	filterFilesWithOrderedPatterns,
-	getClosestBiomeConfigPath,
-	loadRawBiomeConfigFile,
-	resolveExtendsChainGeneric,
-} from "./biomeConfigUtils";
+import { getClosestBiomeConfigPath, loadRawBiomeConfigFile } from "./biomeConfigUtils";
 import type { GitRepo } from "./gitRepo";
 
 /**
@@ -410,6 +407,85 @@ export async function getBiome2FormattedFilesFromDirectory(
 	}
 	const config = await loadBiome2Config(configFile);
 	return getBiome2FormattedFiles(config, directory, gitRepo);
+}
+
+/**
+ * Filters files from git using ordered patterns from Biome 2.x config.
+ * Patterns are processed in declaration order, supporting re-inclusion patterns.
+ *
+ * In Biome 2.x, the 'includes' field uses a unified syntax where:
+ * - Regular patterns (e.g., "src/**") indicate files to include
+ * - Negated patterns (prefixed with `!`, e.g., "!node_modules/**") indicate files to exclude
+ *
+ * Patterns are processed in order, allowing for complex re-inclusion patterns like:
+ * `["**", "!test/**", "test/special/**"]` which means:
+ * 1. Include all files (**)
+ * 2. Exclude test/** (!test/**)
+ * 3. Re-include test/special/** (test/special/**)
+ *
+ * @param orderedPatterns - Array of patterns in declaration order (with ! prefix for exclusions)
+ * @param directory - The repo-root-relative directory to enumerate files from.
+ * @param gitRepo - A GitRepo instance that is used to enumerate files.
+ * @param prefixGlob - Optional function to transform patterns (e.g., prepend "**\/").
+ * @returns Array of absolute paths to files that match the patterns.
+ */
+async function filterFilesWithOrderedPatterns(
+	orderedPatterns: string[],
+	directory: string,
+	gitRepo: GitRepo,
+	prefixGlob?: (glob: string) => string,
+): Promise<string[]> {
+	// Get all files from git (these are already filtered by .gitignore)
+	const allFiles = await gitRepo.getFiles(directory);
+
+	// Apply prefix function if provided
+	const prefixedPatterns = prefixGlob ? orderedPatterns.map(prefixGlob) : orderedPatterns;
+
+	// Find the first exclusion pattern to separate initial includes from the rest
+	const firstExclusionIndex = prefixedPatterns.findIndex((p) => p.startsWith("!"));
+
+	// If no exclusions, just apply includes
+	if (firstExclusionIndex === -1) {
+		const filtered =
+			prefixedPatterns.length > 0 ? multimatch(allFiles, prefixedPatterns) : allFiles;
+		const repoRoot = gitRepo.resolvedRoot;
+		return filtered.map((filePath) => path.resolve(repoRoot, filePath));
+	}
+
+	// Get initial include patterns (before first exclusion)
+	const initialIncludes = prefixedPatterns.slice(0, firstExclusionIndex);
+	const restPatterns = prefixedPatterns.slice(firstExclusionIndex);
+
+	// Start with files matching initial includes (or all files if no initial includes)
+	let result: Set<string>;
+	if (initialIncludes.length > 0) {
+		result = new Set(multimatch(allFiles, initialIncludes));
+	} else {
+		result = new Set(allFiles);
+	}
+
+	// Process remaining patterns in order
+	for (const pattern of restPatterns) {
+		if (pattern.startsWith("!")) {
+			// Exclusion: remove matching files from result
+			const exclusionPattern = pattern.slice(1);
+			const ig = ignore().add([exclusionPattern]);
+			const toRemove = [...result].filter((f) => ig.ignores(f));
+			for (const f of toRemove) {
+				result.delete(f);
+			}
+		} else {
+			// Re-inclusion: add back matching files from original set using multimatch
+			const toAdd = multimatch(allFiles, [pattern]);
+			for (const f of toAdd) {
+				result.add(f);
+			}
+		}
+	}
+
+	// Convert repo root-relative paths to absolute paths
+	const repoRoot = gitRepo.resolvedRoot;
+	return [...result].map((filePath) => path.resolve(repoRoot, filePath));
 }
 
 /**
