+++ build-tools/packages/build-tools/src/test/data/biome2/microsyntax-test/biome.jsonc
