+++ build-tools/packages/build-tools/src/test/data/biome2/microsyntax-test/packages/nested/child/biome.jsonc
